kilometers to miles

+++ KilometersToMiles.java
@@ -3,7 +3,7 @@ public class KilometersToMiles{
 public static void main(String []args){
 Scanner input=new Scanner(System.in);
 System.out.println("Enter the kilometer :");
-double k=input.nextInt();
+double k=input.nextDouble();
 
 System.out.print("Miles is :"+k*0.621371);
 }
